requirements.txt
bcrypt==4.0.1
blinker==1.5
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
pi==0.1.2
protobuf==3.20.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dotenv==0.21.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
